workspace-analysis/pyproject.toml
[tool.poetry]
name = "workspace-analysis"
version = "0.1.0"
description = "A comprehensive set of tools for analyzing the contributions to the Helpful Engineering Slack community"
authors = ["0x2b3bfa0 <0x2b3bfa0+git@googlemail.com>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.7"
cachier = "^1.5.0"
click = "^7.1.2"
matplotlib = "^3.3.4"
numpy = "^1.20.1"
Pillow = "^8.1.2"
slackclient = "^2.9.3"
wordcloud = "^1.8.1"
airtable-python-wrapper = "^0.15.1"
spacy = "^3.0.3"
en_core_web_sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl"}

[tool.poetry.dev-dependencies]
poethepoet = "^0.10.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
